PR Changes: Reformatted output; Readme addition
