+++ .gitignore
@@ -183,6 +183,5 @@ cython_debug/
 nohup.out
 *.tar.gz
 *.pth
-scripts/lens_output.json
-scripts/cleaned_lens_output.json
-scripts/labeled_data.csv
+scripts/lens-cleaner/lens_output.json
+scripts/lens-cleaner/labeled_data.csv
